+++ platform/c2000/inverter/c2000pwmdriver.cpp
@@ -114,7 +114,7 @@ void PwmDriver::EnableACHeatOutput()
  * \param limNeg    Negative over current limit
  * \param limPos    Positive over current limit
  */
-void PwmDriver::SetOverCurrentLimits(int limNeg, int limPos)
+void PwmDriver::SetOverCurrentLimits(int16_t limNeg, int16_t limPos)
 {
 }
 
@@ -154,7 +154,7 @@ void PwmDriver::AcHeat(s32fp ampnom)
  * Set the charge current target
  * \param dc    Target current
  */
-void PwmDriver::SetChargeCurrent(int dc)
+void PwmDriver::SetChargeCurrent(int16_t dc)
 {
 }
 
@@ -162,7 +162,7 @@ void PwmDriver::SetChargeCurrent(int dc)
  * Obtain how many PWM ticks we spend running the main control loop
  * \return Number of ticks
  */
-int PwmDriver::GetCpuLoad()
+int16_t PwmDriver::GetCpuLoad()
 {
     return 100;
 }
